refactor: split diagnostic families
